Fixed bug causing the CheapestOneFree
If the individual item discount succeeds and creates a remainder, said
remainder will not be calculated in the downstream grouped discount.

Fixed by altering the quantity of the item to the remainder. This
required parsing a deepcopy of the items to the offers to ensure it
does not alter the Baskets item quanities.

Added additional offer tests